It occurred that tx history fetching time linearly depends on number of addresses.

Measurements:
Before
3000 transactions + 5 addresses -> 7 sec
3000 transactions + 1000 addresses -> 40 sec

Now
3000 transactions + 5 addresses -> 7 sec
3000 transactions + 1000 addresses -> 12 sec

Current estimated duration for bittrex is 32 sec.

As it was noticed, there is a space for optimizing related request to database, but for now it doesn't play significant role. We will perform this work in master