@domenkozar 2017-10-03
This will also need a quickcheck for regression testing.