@dcoutts 2017-10-03
So `getHistoryLimited` does the following:

 * One DB query to get the `getWalletAccountIds`, which itself is O(n) in the number of accounts in the wallet.
 * N DB queries `getAccountWAddresses`, each of which is O(n * log n) in the number of addresses (because it sorts them).
 * Optionally (if a specific address is specified) we do a linear search within the nacc * naddr number of addresses
 * **Another** call to `getWalletAccountIds` (within `getFullWalletHistory`), repeating the DB query
 * **Another** N DB queries by repeating the calls `getAccountWAddresses`, but this time for every single account in the wallet, rather than just for selected accounts as we did the first time.
 * Then does a topological sort on the transactions in the local mempool, costing O (n * log n) in the size of the local mempool.
 * It then goes through all the mempool transactions comparing their in and out addresses to every single address in every account in the selected wallet, so O (nacc * naddr * nmempool * (avg_in + avg_out)).
 * converts the local history from a list to a DList and back to a list for an extra O(nmempool).
 * Note that currently the local mempool will typically be empty for edge nodes, but if/when we turn on sending transactions to all nodes then the mempool can be large. So this expense is currently small but could be dramatically bigger.
 * converts the block cache from a list to a DList and back to a list for an extra O(nbcache).
 * then we filter the local history costing O(nbcache + log nbcache * nmempool), (whereas this could be done as a set difference).
 * we convert the local history into a DList then convert it back **again**.
 * then we do another DB query to `getWalletPendingTxs`
 * then we do another load of DB queries, O(nbcache + nmempool), by calling `getWalletThTime` on each `c` in `currentHistory`.
 * for **every** entry in the `fullHistory` (which is at least nbcache + nmempool big) we do:
    * execute an STM transaction to get the `getLastKnownHeader`, which means it's not consistent since we do it for every entry in the `fullHistory` so it could change.
    * execute two block DB queries.
    * perform a wallet DB **update** (`addOnlyNewTxMeta`).
    * perform another wallet DB query (`getTxMeta`).
    * perform another wallet DB query (`getPendingTx`).
    * perform another wallet DB query **for each account** in the wallet (`getWalletAddrMetas`)

And this is why we can't have nice things.

What we should do:

 * A **single** DB query, and a single STM query to get all the data we need, so that the data is guaranteed to be consistent. Multiple transactions means non-consistent data.
 * A **pure** function to transform the DB query into the results (pure, so no more DB queries or updates).
 * Analyse the computational complexity of this function.
 * If necessary, maintain caches in the wallet DB and update them on DB updates.